pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "imt-tree-utils"
version = "0.2.7"
authors = [
  { name="Simon Bachhuber", email="simon.bachhuber@fau.de" },
]
description = "Utilities for working with Pytrees"
readme = "readme.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "jax", "jaxlib", "dm-tree"
]

[project.urls]
Homepage = "https://github.com/SimiPixel/tree_utils"
